requirements.txt
openai-whisper>=20240930
soundfile>=0.12.1
numpy>=1.24
